requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.3
marshmallow-sqlalchemy==1.4.2
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
